Fix 64-bit pointer truncation and engine-table cast losses
Several long-standing patterns silently truncate native pointers when
PAWN_CELL_SIZE/sizeof(int)=4 ≠ sizeof(void*)=8:

- BotFunc_FillString computed `(int)ptr - (int)string` to find a fill
  offset; both ptrs are char* in the same buffer, switch to plain
  ptrdiff_t subtraction.
- megahal's strpos did the same `int(pos) - int(start)` cast.
- MAKE_STRING(str) subtracted via int casts; route through intptr_t
  before narrowing back to the engine's int32 string-pool offset.
- BOT_TASK_USE_TELEPORTER stuffed a second edict_t* into CBotTask's int
  m_iInfo slot. Widened m_iInfo / TaskInt() / SetInt to intptr_t so a
  pointer round-trips on 64-bit; consumers that need a real int (Impulse,
  weapon ids, waypoint indices) implicitly narrow as before.
- engine.cpp / dll.cpp pfn debug logs printed edict_t* via `unsigned(e)`
  + `%x`; now use `static_cast<const void*>(e)` + `%p`.

While here:
- Replace `std::_Array_const_iterator<...>` (MSVC-internal type) with
  `auto` so libstdc++ can deduce the actual iterator type.
- Add a few sets of parens in `a && b || c` chains where -Wparentheses
  flagged ambiguous precedence.

Verified on docker-cross builds: i386 / amd64 / aarch64 all link clean,
no behavioral change on i386.

+++ dlls/bot.cpp
@@ -2216,7 +2216,6 @@ bool BotFunc_FillString(char* string, const char* fill_point, const char* fill_w
 		std::memset(before, 0, len);
 		std::memset(after, 0, len);
 
-		//const int start = reinterpret_cast<int>(ptr) - reinterpret_cast<int>(string);
 		const ptrdiff_t start = ptr - string;
 		std::strncpy(before, string, start);
 
@@ -2224,7 +2223,7 @@ bool BotFunc_FillString(char* string, const char* fill_point, const char* fill_w
 		string[start] = 0;
 
 		// Use fill_point_length inside the loop
-		const size_t end = start + fill_point_length;
+		const size_t end = static_cast<size_t>(start) + fill_point_length;
 		std::strncpy(after, &string[end], len - end);
 		after[len - end] = 0;
 
@@ -6492,7 +6491,7 @@ bool CBot::UpdateVisibles()
 	{
 		const int iSched = m_Tasks.GetNewScheduleId();
 
-		AddPriorityTask(CBotTask(BOT_TASK_USE_TELEPORTER, iSched, pTeleporter->v.euser1, reinterpret_cast<int>(pTeleporter), 0,
+		AddPriorityTask(CBotTask(BOT_TASK_USE_TELEPORTER, iSched, pTeleporter->v.euser1, reinterpret_cast<intptr_t>(pTeleporter), 0,
 			pTeleporter->v.origin));
 		AddPriorityTask(CBotTask(BOT_TASK_FIND_PATH, iSched, pTeleporter));
