+++ dlls/util.h
@@ -35,7 +35,10 @@ extern globalvars_t* gpGlobals;
 
 // Use this instead of ALLOC_STRING on constant strings
 #define STRING(offset)		(const char *)(gpGlobals->pStringBase + (int)(offset))
-#define MAKE_STRING(str)	((int)(str) - (int)STRING(0))
+// HL engine string-pool offset is int32. Round-trip through intptr_t so the
+// cast-back to int doesn't lose precision warnings on 64-bit hosts (same
+// arithmetic answer as long as both pointers live inside the same pool).
+#define MAKE_STRING(str)	((int)((intptr_t)(str) - (intptr_t)STRING(0)))
 
 inline edict_t* FIND_ENTITY_BY_CLASSNAME(edict_t* entStart, const char* pszName)
 {
